+++ src/server.ts
@@ -22,7 +22,7 @@ import RoundThumbnailManager from './RoundThumbnailManager';
 import albumRoutes from './routes/album_routes';
 import clubRoutes from './routes/club_routes';
 import memberRoutes from './routes/member_routes';
-import memberMatchRoutes from './routes/member_match_routes';
+import memberInsightsRoutes from './routes/member_insights_routes';
 import roundRoutes from './routes/round_routes';
 import authRoutes from './routes/auth_routes';
 import spotifyRoutes from './routes/spotify_routes';
